smclient: Add ability to connect lately to the session manager
For this, the mode should be set to EGG_SM_CLIENT_MODE_DISABLED early in
the application, and then setting the mode to
EGG_SM_CLIENT_MODE_NO_RESTART or EGG_SM_CLIENT_MODE_NORMAL will make the
registration occur if needed. This is useful for applications that need
to control how the session registration is done, like gnome-panel.

Internally, this means that instead of creating a dummy EggSMClient
object when the mode is EGG_SM_CLIENT_MODE_DISABLED, we directly create
the right object, but we don't connect to the session manager when
parsing the options.

Also, some checks are now done in egg_sm_client_set_mode() so that the
mode cannot be changed in ways that wouldn't work (eg: changing the
state to EGG_SM_CLIENT_MODE_DISABLED while the client has already
connected to the sessio manager).